Fixed some collision issues
Applied a bouncy physical material to the player and enemy prefab.

Used AddRelativeForce instead of Transform.position because the rigidbody component I think does not register the transform.position as a velocity, and bouncy would not work.

I used drag on the rigid to stop the player going too far or fast.

The player, enemy and earth should bouncy off of other.